requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
gunicorn>=20.1.0
python-multipart>=0.0.5
jinja2>=3.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.22.0,<0.23.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
python-dotenv>=0.19.0
